Add scripts to run docker images.
Will run client and server containers. The client script requires that a video be passed into it. Just make sure it is in the same path. There is also a unique issue to using these Docker containers that does not occur when running the rev2.0 library. There seems to be a limit to the number of frames that can be sent. More tests are needed to investigate possible solutions.